Share webhook handlers for Hono parity
Moves webhook request/response mapping behind transport-neutral Request handlers so the control-plane service can mount the route family without importing Next route modules while preserving existing Next behavior.

Constraint: Issue #344 requires Hono/control-plane parity for webhooks while preserving existing Next route response bodies, status codes, auth failures, validation aliases, and tenant scoping.

Rejected: Importing src/app/api/webhooks route modules from services/api | keeps Hono coupled to Next adapter internals and manufactured params.

Confidence: high

Scope-risk: narrow

Directive: Keep future control-plane parity slices on shared handler or core-service seams; do not route Hono through Next App Router modules.

Tested: make check; bunx vitest run tests/webhooks-api-keys-tenant-isolation.test.ts tests/api-auth-date-range-routes.test.ts -t "webhook|control-plane webhooks|API key permission"; make test

Not-tested: Playwright e2e; no dashboard UI changed.
