+++ src/lib/api/webhooks/index.ts
@@ -1,0 +1,7 @@
+export {
+  handleCreateWebhookRequest,
+  handleDeleteWebhookRequest,
+  handleGetWebhookRequest,
+  handleListWebhooksRequest,
+  handleUpdateWebhookRequest,
+} from "./handlers";
